requirements.txt
# Google ADK and Vertex AI
google-genai-adk>=1.0.0
vertexai>=1.70.0
google-cloud-aiplatform>=1.70.0

# Blender Integration
bpy>=4.0.0  # Blender Python API (requires Blender installation)
fake-bpy-module-4.0>=20240604  # For development without Blender

# Core Dependencies
pydantic>=2.9.2
python-dotenv>=1.0.0
pyyaml>=6.0.1
structlog>=24.4.0
tenacity>=8.2.0

# Data Processing
numpy>=1.24.0
pillow>=10.0.0
trimesh>=4.0.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-mock>=3.14.0

# Development
black>=24.8.0
isort>=5.13.0
mypy>=1.11.0